Include OpenType font files in PDF output
Add otffiles=file1,file2,... extra to NEWPAGE command for PDF output.
These OpenType fonts are then available for use in LABEL command and
are included in the PDF output file.

OpenType fonts support Unicode, enabling Unicode to be used directly
in Mapyrus (Type 1 fonts are limited to 256 characters so using a
wide range of characters is not possible).